# cortes-ia-contracts/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortes IA — contrato proposto",
    "version": "0.2.0",
    "description": "Contrato-alvo atualizado pelas decisões v0.2. Não confundir com OpenAPI efetivamente gerado pela API inicial; ver API_IMPLEMENTADA.md."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Placeholder deliberado; substituir pelo host do ambiente."
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "auth_register",
        "summary": "Cadastrar conta",
        "description": "Valida dados; resposta neutra evita enumeração. Grant trial após requisitos de elegibilidade/verificação.",
        "tags": [
          "auth"
        ],
        "security": [],
        "responses": {
          "202": {
            "description": "Operação aceita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "auth_login",
        "summary": "Autenticar ou iniciar MFA",
        "description": "Rate limit, origem/antiforgery, cookie somente após fatores exigidos. Nunca retornar sessão persistente ao localStorage.",
        "tags": [
          "auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "summary": "Revogar sessão",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "account_get",
        "summary": "Consultar perfil mascarado",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "account"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Versão para If-Match."
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "account_patch",
        "summary": "Alterar perfil com step-up em dados sensíveis",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "account"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountPatch"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cortes_session"
      },
      "serviceAuth": {
        "type": "mutualTLS",
        "description": "Workload autenticado no ingresso privado; sem acesso público."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ]
      },
...
      "Ack": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "type": "boolean"
          }
        },
        "required": [
          "accepted"
        ]
      },
      "Register": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+[1-9][0-9]{7,14}$"
          },
          "cpf": {
            "type": "string",
            "pattern": "^[0-9]{11}$"
          },
          "password": {
            "type": "string",
            "format": "password",
            "writeOnly": true
          },
          "termsVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "termsAccepted": {
            "const": true
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "cpf",
          "password",
          "termsVersion",
          "termsAccepted"
        ]
      },
...
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "mfaRequired": {
            "type": "boolean"
          },
          "challengeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "authenticated",
          "mfaRequired"
        ]
      },
...
      "Account": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneMasked": {
            "type": "string"
          },
          "cpfMasked": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "DELETION_PENDING",
              "DELETED"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phoneMasked",
          "cpfMasked",
          "status",
          "emailVerified",
          "phoneVerified",
          "version"
        ]
      },
      "AccountPatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": []
      },
...
    }
  }
}